💫 (#9) - Move user-related project queries to User repository
