+++ src/shared/entities/user/user.repository.ts
@@ -1,6 +1,7 @@
-import { AbstractRepository, EntityRepository } from 'typeorm';
+import { AbstractRepository, Brackets, EntityRepository } from 'typeorm';
 import { QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity';
 import { v4 } from 'uuid';
+import { Project } from '../project/project.entity';
 import { User } from './user.entity';
 
 @EntityRepository(User)
@@ -50,5 +51,107 @@ export class UserRepository extends AbstractRepository<User> {
       .set({ deleted: true })
       .where('uuid = :uuid', { uuid })
       .execute();
+  }
+
+  async getProjectsOfUser(
+    userId: number,
+    isMine: boolean,
+    page: number,
+    limit: number,
+  ) {
+    const qb = this.createQueryBuilder('project')
+      .select()
+      .from(Project, 'project')
+      .leftJoinAndSelect('project.members', 'members')
+      .leftJoinAndSelect('members.userId', 'userId')
+      .where('members.userId = :userId', { userId })
+      .orderBy('project.createdAt', 'DESC')
+      .take(limit)
+      .skip((page - 1) * limit);
+    if (!isMine)
+      qb.leftJoinAndSelect('project.status', 'status').andWhere(
+        'report.isReportAccepted = true',
+      );
+
+    return qb.getManyAndCount();
+  }
+
+  async getPendingProjects(userId?: number) {
+    const qb = this.createQueryBuilder('project')
+      .select()
+      .from(Project, 'project')
+      .leftJoinAndSelect('project.status', 'status')
+      .leftJoinAndSelect('project.members', 'members')
+      .where(
+        new Brackets((qb) => {
+          qb.where(
+            new Brackets((qb) => {
+              qb.where('status.isPlanSubmitted = true').andWhere(
+                new Brackets((qb) => {
+                  qb.where('status.isPlanAccepted IS NULL').orWhere(
+                    'status.isPlanAccepted = false',
+                  );
+                }),
+              );
+            }),
+          );
+          qb.orWhere(
+            new Brackets((qb) => {
+              qb.where('status.isReportSubmitted = true').andWhere(
+                new Brackets((qb) => {
+                  qb.where('status.isReportAccepted IS NULL').orWhere(
+                    'status.isReportAccepted = false',
+                  );
+                }),
+              );
+            }),
+          );
+        }),
+      )
+      .take(4);
+    if (userId) qb.andWhere('members.userId = :userId', { userId });
+
+    return qb.getManyAndCount();
+  }
+
+  async getReports(
+    userId: number,
+    page: number,
+    limit: number,
+    type: boolean | null,
+  ) {
+    return this.createQueryBuilder('project')
+      .select()
+      .from(Project, 'project')
+      .leftJoinAndSelect('project.status', 'status')
+      .leftJoinAndSelect('project.members', 'members')
+      .where(
+        new Brackets((qb) => {
+          qb.where(
+            new Brackets((qb) => {
+              qb.where('status.isPlanSubmitted = true').andWhere(
+                type !== null
+                  ? 'status.isPlanAccepted = :type'
+                  : 'status.isPlanAccepted IS NULL',
+                { type },
+              );
+            }),
+          );
+          qb.orWhere(
+            new Brackets((qb) => {
+              qb.where('status.isReportSubmitted = true').andWhere(
+                type !== null
+                  ? 'status.isReportAccepted = :type'
+                  : 'status.isReportAccepted IS NULL',
+                { type },
+              );
+            }),
+          );
+        }),
+      )
+      .andWhere('members.userId = :userId', { userId })
+      .take(limit)
+      .skip(limit * (page - 1))
+      .getManyAndCount();
   }
 }
